requirements.txt
gradio
numpy
pillow
scikit-learn